move recentSearches and Modifiers into Initial component
